More reliable way of detecting blocks not participating in a merge
spv::Block already stores predecessor information. This is more reliable
and future-proof to not only skip the header in if/then/else (cannot
reach merge block from header if else is present) but also to skip
cases where the then/else blocks are break or continue blocks and thus
never reach the selection merge block.