Introduced concept of target with command line argument support
Targets are now a proper thing. The FrontendAction no longer has hard
coded GLSL or SPIR-V references. The AST Consumer now instead uses the
ModuleBuilder as provided by the Target.

Targets support a plugin-like architecture and are selectable with the
-target command line option. They can have their own private command
line arguments and categories to drive code generation. The builtin
targets are statically added in the lava executable.